+++ client/src/store/reducers/theme.ts
@@ -1,20 +1,30 @@
 import { Action } from '../actions';
 import { ActionType } from '../action-types';
 import { Theme } from '../../interfaces/Theme';
+import { arrayPartition, parsePABToTheme } from '../../utility';
 
 interface ThemeState {
-  theme: Theme;
+  activeTheme: Theme;
+  themes: Theme[];
+  userThemes: Theme[];
+  themeInEdit: Theme | null;
 }
 
+const savedTheme = localStorage.theme
+  ? parsePABToTheme(localStorage.theme)
+  : parsePABToTheme('#effbff;#6ee2ff;#242b33');
+
 const initialState: ThemeState = {
-  theme: {
-    name: 'tron',
+  activeTheme: {
+    name: 'main',
+    isCustom: false,
     colors: {
-      background: '#242B33',
-      primary: '#EFFBFF',
-      accent: '#6EE2FF',
+      ...savedTheme,
     },
   },
+  themes: [],
+  userThemes: [],
+  themeInEdit: null,
 };
 
 export const themeReducer = (
@@ -22,8 +32,56 @@ export const themeReducer = (
   action: Action
 ): ThemeState => {
   switch (action.type) {
-    case ActionType.setTheme:
-      return { theme: action.payload };
+    case ActionType.setTheme: {
+      return {
+        ...state,
+        activeTheme: {
+          ...state.activeTheme,
+          colors: action.payload,
+        },
+      };
+    }
+
+    case ActionType.fetchThemes: {
+      const [themes, userThemes] = arrayPartition<Theme>(
+        action.payload,
+        (e) => !e.isCustom
+      );
+
+      return {
+        ...state,
+        themes,
+        userThemes,
+      };
+    }
+
+    case ActionType.addTheme: {
+      return {
+        ...state,
+        userThemes: [...state.userThemes, action.payload],
+      };
+    }
+
+    case ActionType.deleteTheme: {
+      return {
+        ...state,
+        userThemes: action.payload,
+      };
+    }
+
+    case ActionType.editTheme: {
+      return {
+        ...state,
+        themeInEdit: action.payload,
+      };
+    }
+
+    case ActionType.updateTheme: {
+      return {
+        ...state,
+        userThemes: action.payload,
+      };
+    }
 
     default:
       return state;