+++ controllers/themes/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  getThemes: require('./getThemes'),
+  addTheme: require('./addTheme'),
+  deleteTheme: require('./deleteTheme'),
+  updateTheme: require('./updateTheme'),
+};